# kompello-web/app/lib/api/kompello/models/Password.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kompello Server API
 * Kompello API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Password
 */
export interface Password {
    /**
     * 
     * @type {string}
     * @memberof Password
     */
    password: string;
}

/**
 * Check if a given object implements the Password interface.
 */
export function instanceOfPassword(value: object): value is Password {
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function PasswordFromJSON(json: any): Password {
    return PasswordFromJSONTyped(json, false);
}

export function PasswordFromJSONTyped(json: any, ignoreDiscriminator: boolean): Password {
    if (json == null) {
        return json;
    }
    return {
        
        'password': json['password'],
    };
}

export function PasswordToJSON(json: any): Password {
    return PasswordToJSONTyped(json, false);
}

export function PasswordToJSONTyped(value?: Password | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'password': value['password'],
    };
}