# kompello-web/openapi/kompello.yaml
openapi: 3.0.3
info:
  title: Kompello Server API
  version: 1.0.0
  description: Kompello API Documentation
  contact:
    name: 'Github: Kompello'
    url: https://github.com/KBurgTech/kompello
  license:
    name: 'Licence: MIT'
    url: https://github.com/KBurgTech/kompello/blob/main/LICENSE
paths:
...
  /api/users/{uuid}/set_password/:
    post:
      operationId: users_set_password
      description: Change a users password
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Password'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Password'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      responses:
        '200':
          description: No response body
        '403':
          description: No response body
  /api/users/me/:
    get:
      operationId: users_me
      description: Gets the currently logged in user
      tags:
      - users
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
components:
  schemas:
...
    Password:
      type: object
      properties:
        password:
          type: string
      required:
      - password
...
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        created_on:
          type: string
          format: date-time
          readOnly: true
        modified_on:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_on
      - email
      - modified_on
      - uuid
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid